requirements.txt
sox
python-slugify